Organise steps into packages
